+++ src/services/userService.ts
@@ -15,11 +15,11 @@ export const userService = {
     return apiService.post(process.env.REACT_APP_USER_DISABLE_ENDPOINT?.replace('{target_username}', userId) || '', {});
   },
 
-  async createUser(userData: { given_name: string; last_name: string; username: string; email: string; phone_number: string; temporary_password: string }): Promise<ApiResponse> {
+  async createUser(userData: { given_name: string; family_name: string; username: string; email: string; phone_number: string; temporary_password: string; role: string }): Promise<ApiResponse> {
     return apiService.post(process.env.REACT_APP_USER_CREATE_ENDPOINT || '', userData);
   },
 
-  async updateUser(username: string, userData: Partial<{ given_name: string; family_name: string; email: string; phone_number: string }>): Promise<ApiResponse> {
+  async updateUser(username: string, userData: Partial<{ given_name: string; family_name: string; email: string; phone_number: string; role: string }>): Promise<ApiResponse> {
     return apiService.put(process.env.REACT_APP_USER_UPDATE_ENDPOINT?.replace('{target_username}', username) || ``, userData);
   },